ANDROID: rockchip: Remove the boot partition
We do not plan to use the boot partition at the moment, so to eliminate
any confusion, and to unbreak the "bootable partition" feature in the
default U-Boot environment, remove the boot partition.

As part of this change, we also move the 'legacy BIOS bootable' flag
from the removed boot partition to the rootfs partition. This means the
rootfs burned to the device must have a valid /boot directory with the
boot.scr file.

Change-Id: I1d800385cfacd22a812ee3749a50f4a2b899cf85
Signed-off-by: Alistair Delva <adelva@google.com>